Apply patch files
This code is a mess but right now DOES produce a working executable using the current version of
the Mod API exmaple; it calls into the custom code section and then immediately returns back.
This works by making a temporary collection of sections for each patch file and processing
relocations on it using a symbol table shared between patch files and modfiles.

This code desperately needs to be refactored to be more readable, reliable and testable. There
are likely MANY edge cases that will cause malformed XBEs or panic outright.