Linker proof of concept. (#3)
* Begin support for multiple object files

This code currently is very buggy and panics with the current test. The issue appears to be
with the symbol table generation but there are likely bugs all over the place.

* Fix panic issue (code still sucks)

Fix mbss section flags

Fix wrong length being used for file start offset in a section

Look at correct section to update for relocations (Instead of trying to apply
the relocation to the section the referenced symbol is defined in)

* Test SectionInProgress::add_bytes

* Fix clippy lints

* Apply patch files

This code is a mess but right now DOES produce a working executable using the current version of
the Mod API exmaple; it calls into the custom code section and then immediately returns back.
This works by making a temporary collection of sections for each patch file and processing
relocations on it using a symbol table shared between patch files and modfiles.

This code desperately needs to be refactored to be more readable, reliable and testable. There
are likely MANY edge cases that will cause malformed XBEs or panic outright.